chore: Move sharing_sites into video_configuration app
- https://github.com/openedx/edx-platform/issues/37456
