+++ xmodule/video_block/video_block.py
@@ -58,7 +58,7 @@ from xmodule.x_module import (
 )
 from xmodule.xml_block import XmlMixin, deserialize_field, is_pointer_tag, name_to_pathname
 from .bumper_utils import bumperize
-from .sharing_sites import sharing_sites_info_for_video
+from openedx.core.djangoapps.video_config.sharing_sites import sharing_sites_info_for_video
 from .transcripts_utils import (
     Transcript,
     VideoTranscriptsMixin,
